✅Del tester opp i flere filer og reduser antall tester til kun de mest relevante
